+++ pagegraph/graph/element.py
@@ -1,8 +1,9 @@
+from __future__ import annotations
+
 from abc import ABC
 from enum import Enum
 import sys
 from typing import Any, TYPE_CHECKING, Union
-
 
 if TYPE_CHECKING:
     from pagegraph.graph import PageGraph
@@ -18,13 +19,27 @@ class PageGraphElement(ABC):
         # attributes that correspond to node and edge attributes.
 
     # Class properties
+
+    # Inheriting classes can define this property (at the
+    # class level) to include a simple description of the node or edge's
+    # attributes in different reports and summaries. Inheriting classes
+    # that implement should define a dict, mapping "descriptive name of the
+    # value, for use in a JSON dict" to "name of the method to call to get
+    # that value".
     summary_methods: Union[dict[str, str], None] = None
 
     # Instance properties
-    pg: "PageGraph"
-    _id: "PageGraphId"
 
-    def __init__(self, graph: "PageGraph", pg_id: "PageGraphId"):
+    # Reference to the PageGraph instance that this graph element is a member
+    # of (i.e., the graph this node or edge is a member of.)
+    pg: PageGraph
+
+    # The ID of this element in the graph, as defined by the GraphML spec.
+    # This will either be "e<int>" or "n<int>", for edges and notes,
+    # respectively.
+    _id: PageGraphId
+
+    def __init__(self, graph: PageGraph, pg_id: PageGraphId):
         self.pg = graph
         self._id = pg_id
 
@@ -34,11 +49,11 @@ class PageGraphElement(ABC):
     def id(self) -> int:
         return int(self._id[1:])
 
-    def pg_id(self) -> "PageGraphId":
+    def pg_id(self) -> PageGraphId:
         return self._id
 
-    def summary_fields(self) -> "ElementSummary":
-        summary: dict[str, "JSONAble"] = {}
+    def summary_fields(self) -> ElementSummary:
+        summary: dict[str, JSONAble] = {}
         needle_class = self.__class__
         while needle_class != object and needle_class is not ABC:
             class_summary_methods = needle_class.summary_methods
@@ -54,18 +69,29 @@ class PageGraphElement(ABC):
         return summary
 
     def validate(self) -> bool:
+        # Inheriting classes should implement this method to perform correctness
+        # checks and similar activities. This is intended as a check that
+        # the `pagegraph-query` code correctly understands and abstracts the
+        # pagegraph-generated GraphML data.
         raise NotImplementedError()
 
     def data(self) -> dict[str, Any]:
-        raise NotImplementedError("Child class must implement 'data'")
+        raise NotImplementedError()
 
     def timestamp(self) -> int:
         return int(self.data()[self.__class__.RawAttrs.TIMESTAMP.value])
 
     def describe(self) -> str:
-        raise NotImplementedError("Child class must implement 'describe'")
+        # Inheriting classes must implement this method. Its used in debugging
+        # and similar situations to describe the node or edge in STDERR
+        # and equiv situations.
+        raise NotImplementedError()
 
     def build_caches(self) -> None:
+        # Inheriting classes should implement this to perform any
+        # expensive, one-time operations. This method is guaranteed to be
+        # called after the parent PageGraph graph element is initialized, but
+        # before any other, query-related methods are called.
         pass
 
     def throw(self, desc: str) -> None: