+++ pagegraph/graph/node/cookie_jar.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from typing import Optional
 
 from pagegraph.graph.node.storage_area import StorageAreaNode
@@ -5,5 +7,5 @@ from pagegraph.graph.node.storage_area import StorageAreaNode
 
 class CookieJarNode(StorageAreaNode):
 
-    def as_cookie_jar_node(self) -> Optional["CookieJarNode"]:
+    def as_cookie_jar_node(self) -> Optional[CookieJarNode]:
         return self