+++ pagegraph/graph/node/local_storage.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from typing import Optional
 
 from pagegraph.graph.node.storage_area import StorageAreaNode
@@ -5,5 +7,5 @@ from pagegraph.graph.node.storage_area import StorageAreaNode
 
 class LocalStorageNode(StorageAreaNode):
 
-    def as_local_storage_node(self) -> Optional["LocalStorageNode"]:
+    def as_local_storage_node(self) -> Optional[LocalStorageNode]:
         return self