+++ pagegraph/graph/node/web_api.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from typing import Optional
 
 from pagegraph.graph.edge import Edge
@@ -13,5 +15,5 @@ class WebAPINode(JSStructureNode):
         Edge.Types.JS_RESULT
     ]
 
-    def as_web_api_node(self) -> Optional["WebAPINode"]:
+    def as_web_api_node(self) -> Optional[WebAPINode]:
         return self